+++ mathematica/surprising-polynomial-identities-classical-interpolation.m
@@ -9,12 +9,14 @@ PrintTriangleA::usage="PrintTriangleA[m] prints triangle of coefficients A for g
 BivariateSumT::usage="Defines bivariate sum T(m,n,k)."
 BackwardRecurrenceForT::usage="Validates the backward recurrence for the bivariate sum T(m,n,k)."
 ForwardRecurrenceForT::usage="Validates the forward recurrence for the bivariate sum T(m,n,k)."
+CentralRecurrenceForT::usage="Validates the central recurrence for the bivariate sum T(m,n,k)."
 OddPowerDecomposition::usage="Validates the identity Odd power decomposition."
 OddPowerDecompositionMMinus1::usage="Validates the identity Odd power decomposition m-1."
 OddPowerBinomialForm::usage="Validates the identity Odd power binomial form shifted."
 OddPowerBinomialFormShifted::usage="Validates the identity Odd power binomial form."
 TableFormBackwardRecurrenceForT::usage="Prints the backward recurrence for Tm in the form of triangle."
 TableFormForwardRecurrenceForT::usage="Prints the forward recurrence for Tm in the form of triangle."
+TableFormCentralRecurrenceForT::usage="Prints the central recurrence for Tm in the form of triangle."
 TableFormBivariateSumT::usage="Prints Tm in the form of triangle."
 
 Begin["`Private`"]
@@ -31,13 +33,15 @@ OddPowerIdentitySimplified[n_, m_]:= Expand[Sum[Sum[A[m,r] * k^r * (n-k)^r, {k, 
 PrintTriangleA[m_]:= TableForm[Table[A[n, k], {n, 0, m}, {k, 0, n}], TableAlignments -> Left];
 BivariateSumT[m_, n_, k_]:= Sum[A[m, r] * k^r (n-k)^r, {r, 0, m}];
 BackwardRecurrenceForT[m_, n_, k_]:= Sum[(-1)^(t-1) * Binomial[m+1, t] * BivariateSumT[m, n-t, k], {t, 1, m+1}];
-ForwardRecurrenceForT[m_, n_, k_]:= Sum[(-1)^(m-t) * Binomial[m+1, t] * BivariateSumT[m, n+t, k], {t, 1, m+1}];
+ForwardRecurrenceForT[m_, n_, k_]:= Sum[(-1)^(t+1) * Binomial[m+1, t] * BivariateSumT[m, n+t, k], {t, 1, m+1}];
+CentralRecurrenceForT[m_, n_, k_]:= Sum[(-1)^(t+1) * Binomial[m+1, t] * BivariateSumT[m, n+(m/2)-t, k], {t, 1, m+1}];
 OddPowerDecomposition[n_, m_]:= Sum[Sum[(-1)^(t-1) * Binomial[m+1, t] * BivariateSumT[m, n-t, k], {t, 1, m+1}], {k, 1, n}];
 OddPowerDecompositionMMinus1[n_, m_]:= Sum[Sum[(-1)^(t-1) * Binomial[m, t] * BivariateSumT[m-1, n-t, k], {t, 1, m}], {k, 1, n}];
 OddPowerBinomialForm[m_, n_, a_] := Sum[A[m, r]* Sum[(k - a)^r * (n - a - k)^r, {k, a+1, n - a}], {r, 0, m}];
 OddPowerBinomialFormShifted[m_, n_, a_] := Sum[A[m, r]* Sum[(k - a)^r * (n - a - k)^r, {k, a, n - a - 1}], {r, 0, m}];
 TableFormBackwardRecurrenceForT[m_, rows_]:= TableForm[Table[BackwardRecurrenceForT[m, n, k], {n, 0, rows}, {k, 0, n}], TableAlignments -> Left];
 TableFormForwardRecurrenceForT[m_, rows_]:= TableForm[Table[ForwardRecurrenceForT[m, n, k], {n, 0, rows}, {k, 0, n}], TableAlignments -> Left];
+TableFormCentralRecurrenceForT[m_, rows_]:= TableForm[Table[CentralRecurrenceForT[m, n, k], {n, -m/2, rows}, {k, 0, n+m/2}], TableAlignments -> Left];
 TableFormBivariateSumT[m_, rows_]:=TableForm[Table[BivariateSumT[m, n, k], {n, 0, rows}, {k, 0, n}], TableAlignments -> Left];
 
 End[ ]