+++ CHANGELOG.md
@@ -26,3 +26,4 @@ and this project adheres to [Semantic Versioning v2.0.0](https://semver.org/spec
 - Add labels for main results
 - Update files structure
 - Add forward and backward recurrences for Tm
+- Add central recurrence for Tm